refactor(abacus): comprehensive provider overhaul addressing review feedback
Split AbacusUsageFetcher.swift (396 lines) into 4 files matching
the Kimi/Perplexity provider pattern:
- AbacusCookieImporter.swift — browser cookie extraction
- AbacusUsageFetcher.swift — API fetching logic
- AbacusUsageSnapshot.swift — data model + conversion
- AbacusUsageError.swift — error types

Bug fixes:
- Pin NumberFormatter locale to en_US (was locale-dependent)
- Cached cookie parse failures now clear cache and fall through
  to fresh browser import instead of aborting
- Session loop retries on all recoverable errors (auth + parse),
  not just auth — prevents one stale source blocking valid ones
- Preserve last error on loop exhaustion instead of generic
  .noSessionCookie
- Log billing fetch failures instead of silently discarding
- Truncate HTTP response body in error messages (security)
- Capture JSON parse error details instead of using try?
- Throw parseFailed when critical credit fields missing instead
  of returning hollow 0% snapshot

Architecture alignment:
- Add structured CodexBarLog logging (abacusCookie, abacusUsage)
- Use cookieImportCandidates(using:) for browser filtering
- Accept BrowserDetection parameter for testability
- Add hasSession() convenience for isAvailable() check
- Improve isAvailable() to probe for actual session existence
- Add import AppKit (was implicit via SwiftUI)
