+++ Sources/CodexBar/Providers/Abacus/AbacusProviderImplementation.swift
@@ -1,3 +1,4 @@
+import AppKit
 import CodexBarCore
 import CodexBarMacroSupport
 import Foundation
